Move delta debug bits into their own subfolder
